requirements.txt
langchain
langchain-groq
langchain-core
langchain-community
faiss-cpu
sentence-transformers
pypdf
streamlit
python-dotenv
